migrate startWatchingExternalVideo action
