+++ bigbluebutton-html5/imports/api/external-videos/server/methods.js
@@ -1,10 +1,8 @@
 import { Meteor } from 'meteor/meteor';
-import startWatchingExternalVideo from './methods/startWatchingExternalVideo';
 import stopWatchingExternalVideo from './methods/stopWatchingExternalVideo';
 import emitExternalVideoEvent from './methods/emitExternalVideoEvent';
 
 Meteor.methods({
-  startWatchingExternalVideo,
   stopWatchingExternalVideo,
   emitExternalVideoEvent,
 });
