--- bigbluebutton-html5/imports/api/external-videos/server/methods/startWatchingExternalVideo.js
@@ -1,26 +1,0 @@
-import { check } from 'meteor/check';
-import Logger from '/imports/startup/server/logger';
-import RedisPubSub from '/imports/startup/server/redis';
-import { extractCredentials } from '/imports/api/common/server/helpers';
-
-export default function startWatchingExternalVideo(externalVideoUrl) {
-  const REDIS_CONFIG = Meteor.settings.private.redis;
-  const CHANNEL = REDIS_CONFIG.channels.toAkkaApps;
-  const EVENT_NAME = 'StartExternalVideoPubMsg';
-
-  try {
-    const { meetingId, requesterUserId } = extractCredentials(this.userId);
-
-    check(meetingId, String);
-    check(requesterUserId, String);
-    check(externalVideoUrl, String);
-
-    const payload = { externalVideoUrl };
-
-    Logger.info(`User ${requesterUserId} sharing an external video ${externalVideoUrl} for meeting ${meetingId}`);
-
-    RedisPubSub.publishUserMessage(CHANNEL, EVENT_NAME, meetingId, requesterUserId, payload);
-  } catch (error) {
-    Logger.error(`Error on sharing an external video for meeting ${meetingId}: ${error}`);
-  }
-}
