+++ bigbluebutton-html5/imports/ui/components/external-video-player/external-video-player-graphql/modal/service.ts
@@ -8,22 +8,6 @@ export const stopWatching = () => {
   makeCall('stopWatchingExternalVideo');
 };
 
-export const startWatching = (url: string) => {
-  let externalVideoUrl = url;
-
-  if (YOUTUBE_SHORTS_REGEX.test(url)) {
-    const shortsUrl = url.replace('shorts/', 'watch?v=');
-    externalVideoUrl = shortsUrl;
-  } else if (PANOPTO_MATCH_URL.test(url)) {
-    const m = url.match(PANOPTO_MATCH_URL);
-    if (m && m.length >= 4) {
-      externalVideoUrl = `https://${m[1]}/Podcast/Social/${m[3]}.mp4`;
-    }
-  }
-
-  makeCall('startWatchingExternalVideo', externalVideoUrl);
-};
-
 export const isUrlValid = (url: string) => {
   if (YOUTUBE_SHORTS_REGEX.test(url)) {
     const shortsUrl = url.replace('shorts/', 'watch?v=');
@@ -36,5 +20,4 @@ export const isUrlValid = (url: string) => {
 
 export default {
   stopWatching,
-  startWatching,
 };
